requirements.txt
Django>=3.2
djangorestframework>=3.12
scikit-learn>=0.24
pandas>=1.2
numpy>=1.19